Reorganize files for /actions and /models, remove server action declaration for GET requests
