+++ ui/src/components/ui/select.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import * as SelectPrimitive from '@radix-ui/react-select';
 import { Check, ChevronDown, ChevronUp } from 'lucide-react';
 
-import { cn } from '@/components/ui/utils';
+import { cn } from '@/lib/utils';
 
 const Select = SelectPrimitive.Root;
 
@@ -42,10 +42,7 @@ const SelectScrollUpButton = React.forwardRef<
 >(({ className, ...props }, ref) => (
     <SelectPrimitive.ScrollUpButton
         ref={ref}
-        className={cn(
-            'flex cursor-default items-center justify-center py-1',
-            className
-        )}
+        className={cn('flex cursor-default items-center justify-center py-1', className)}
         {...props}
     >
         <ChevronUp className="h-4 w-4" />
@@ -59,17 +56,13 @@ const SelectScrollDownButton = React.forwardRef<
 >(({ className, ...props }, ref) => (
     <SelectPrimitive.ScrollDownButton
         ref={ref}
-        className={cn(
-            'flex cursor-default items-center justify-center py-1',
-            className
-        )}
+        className={cn('flex cursor-default items-center justify-center py-1', className)}
         {...props}
     >
         <ChevronDown className="h-4 w-4" />
     </SelectPrimitive.ScrollDownButton>
 ));
-SelectScrollDownButton.displayName =
-    SelectPrimitive.ScrollDownButton.displayName;
+SelectScrollDownButton.displayName = SelectPrimitive.ScrollDownButton.displayName;
 
 const SelectContent = React.forwardRef<
     React.ElementRef<typeof SelectPrimitive.Content>,
@@ -113,11 +106,7 @@ const SelectLabel = React.forwardRef<
     React.ElementRef<typeof SelectPrimitive.Label>,
     React.ComponentPropsWithoutRef<typeof SelectPrimitive.Label>
 >(({ className, ...props }, ref) => (
-    <SelectPrimitive.Label
-        ref={ref}
-        className={cn('py-1.5 pl-8 pr-2 text-base font-semibold', className)}
-        {...props}
-    />
+    <SelectPrimitive.Label ref={ref} className={cn('py-1.5 pl-8 pr-2 text-base font-semibold', className)} {...props} />
 ));
 SelectLabel.displayName = SelectPrimitive.Label.displayName;
 
@@ -153,10 +142,7 @@ const SelectSeparator = React.forwardRef<
 >(({ className, ...props }, ref) => (
     <SelectPrimitive.Separator
         ref={ref}
-        className={cn(
-            '-mx-1 my-1 h-px bg-slate-100 dark:bg-slate-800',
-            className
-        )}
+        className={cn('-mx-1 my-1 h-px bg-slate-100 dark:bg-slate-800', className)}
         {...props}
     />
 ));
