+++ ui/tests/app/(home)/_components/announcements.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import Announcements from '@/app/(home)/_components/announcements';
-import * as GroupingsApi from '@/actions/groupings-api';
+import * as Fetchers from '@/lib/fetchers';
 
-jest.mock('@/actions/groupings-api');
+jest.mock('@/lib/fetchers');
 
 const message = 'test announcement';
 const message1 = 'test1 announcement';
@@ -33,7 +33,7 @@ const announcements = {
 
 describe('Announcements', () => {
     it('renders announcement correctly', async () => {
-        jest.spyOn(GroupingsApi, 'getAnnouncements').mockResolvedValue(announcements);
+        jest.spyOn(Fetchers, 'getAnnouncements').mockResolvedValue(announcements);
         render(await Announcements());
 
         expect(screen.getAllByLabelText('icon')).toHaveLength(2);
